requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.5
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.21.6
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sklearn==0.0.post5
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
